Add files via upload
My first project is based on analyzing the the distribution of categorial or continuous variable such as the distribution of ages or genders in population and my second project is based on analyzing the sentiment patterns in social media data to understand public opinion and attitudes towards specific topic. The internship was offered by @Prodigy Info Tech.